Weather Station: Added watchdog timer Added ping checks Added signalError when not compiled in DEBUG Messaging init will retry until no error code is returned Added light signal when sending message not in DEBUG Fixed bug when both relaying and handling a message (using seek) Fixed typo in getWindSpeedByte
Receiver:
Added pings
Improved console redirect
Can now send commands to the modem

+++ ArduinoWeatherStation/MessagingCommon.cpp
@@ -63,7 +63,7 @@ MESSAGE_RESULT MessageSource::readBytes(byte* dest, size_t dataLen)
   return MESSAGE_OK;
 }
 
-size_t MessageSource::getCurrentLocation()
+int MessageSource::getCurrentLocation()
 {
   return m_iCurrentLocation;
 }
